Checkpoint: Created "Foundry", but reactivity is broken
